Website + Query class
Added a Query class which I will start to implement where queries are needed. This class should help with formatting queries and make them more readable for everyone. The website was a version of one of the earlier archives which I decided to add in this repo. Will probably not work on the website for a while though.